chore: cleanup and move test matrix tests to the appropriate cmd/xxxx_test.go
